Allow to update a user with "unchanged" fields
When the request for a user update contains the current username shoreline would reject as it interpret it as a duplicate  (which is not).
From the client side it's often practical to send the full user object to shoreline (username, email, roles...) even if only one of those fields have changed.